requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
django-environ==0.9.0